_archive/agent-x/requirements.txt
tweepy>=4.14.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
schedule>=1.2.1
# Optional vector store
chromadb>=0.5.0
